Group tests on the same folder
