+++ tests/test_permutation_collection.py
